+++ src/i18n/zh.ts
@@ -177,6 +177,11 @@ const zh: LANG_MESSAGE = {
   displayAllFeatures: '显示所有功能',
   displayAllFeaturesTip:
     '显示所有功能，包括sing-box官方版本不支持的功能，如果您使用了fork版本的sing-box支持其中的某些功能，可以尝试启用',
+  blurIntensity: '毛玻璃强度',
+  scrollAnimationEffect: '滚动动画效果',
+  importFromFile: '从文件导入',
+  importFromUrl: '从 URL 导入',
+  sync: '同步',
 }
 
 export default zh