+++ src/views/HomePage.vue
@@ -8,7 +8,7 @@
       >
         <div
           v-if="ctrlsMap[route.name as string] && isSidebarCollapsed"
-          class="bg-base-100 w-full"
+          class="bg-base-100 ctrls-bar w-full"
         >
           <component
             :is="ctrlsMap[route.name as string]"
@@ -32,7 +32,7 @@
         </div>
         <template v-if="isMiddleScreen">
           <div
-            class="shrink-0"
+            class="nav-bar shrink-0"
             :class="isPWA ? 'h-[5.5rem]' : 'h-14'"
           />
           <div
@@ -47,7 +47,7 @@
             >
               <component
                 :is="ROUTE_ICON_MAP[r]"
-                class="size-[1.2em]"
+                class="size-5"
               />
               <span class="dock-label">
                 {{ $t(r) }}